requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-elevenlabs
livekit-plugins-deepgram
livekit-plugins-silero
python-dotenv
aiohttp
pytest
twilio
quart
hypercorn
livekit-api